thinkwise-backend/app/services/vector_db_service.py
from typing import List, Dict, Any
import numpy as np

class VectorDBService:
    def __init__(self):
        # TODO: 실제 벡터 DB 연동 구현 (Pinecone, Weaviate, Chroma 등)
        self.documents = []
        self.embeddings = []
        
    async def store_document(self, content: str, metadata: Dict[str, Any] = None) -> str:
        """문서를 벡터 DB에 저장합니다."""
        # TODO: 실제 벡터 DB 저장 로직 구현
        doc_id = f"doc_{len(self.documents)}"
        self.documents.append({
            "id": doc_id,
            "content": content,
            "metadata": metadata or {}
        })
        return doc_id
    
    async def search_similar(self, query: str, top_k: int = 5) -> List[Dict[str, Any]]:
        """쿼리와 유사한 문서를 검색합니다."""
        # TODO: 실제 벡터 유사도 검색 로직 구현
        # 현재는 간단한 키워드 매칭으로 시뮬레이션
        
        results = []
        query_lower = query.lower()
        
        for doc in self.documents:
            content_lower = doc["content"].lower()
            # 간단한 키워드 매칭 점수 계산
            score = 0
            for word in query_lower.split():
                if word in content_lower:
                    score += 1
            
            if score > 0:
                results.append({
                    "id": doc["id"],
                    "content": doc["content"],
                    "metadata": doc["metadata"],
                    "score": score
                })
        
        # 점수 순으로 정렬하고 top_k개 반환
        results.sort(key=lambda x: x["score"], reverse=True)
        return results[:top_k]
    
    async def get_document(self, doc_id: str) -> Dict[str, Any]:
        """문서 ID로 문서를 조회합니다."""
        for doc in self.documents:
            if doc["id"] == doc_id:
                return doc
        return None
    
    async def delete_document(self, doc_id: str) -> bool:
        """문서를 삭제합니다."""
        for i, doc in enumerate(self.documents):
            if doc["id"] == doc_id:
                del self.documents[i]
                return True
        return False
    
    async def update_document(self, doc_id: str, content: str, metadata: Dict[str, Any] = None) -> bool:
        """문서를 업데이트합니다."""
        for doc in self.documents:
            if doc["id"] == doc_id:
                doc["content"] = content
                if metadata:
                    doc["metadata"].update(metadata)
                return True
        return False
